requirements.txt
streamlit==1.44.1
Pillow==11.2.1
requests==2.32.3
google-generativeai==0.8.4
requests==2.32.3
google-api-python-client==2.167.0 
dotenv